BREAKING CHANGE: Simplify API
Allow race to return a branded Cancellation instead of always throwing
it, so that the result can be just checked with isCancellation(result)
or thrown with throwIfCancelled(result).

Make the isCancelled check an internal detail we brand with the oneshot
method and wrap a non branded task inside of newRaceCancellation. This
removes having to pass a isCancelled function and makes the API just
newRaceCancellation(cancellation: () => PromiseLike<unknown>).

The primary API is using the
`const [raceCancellation, cancel] = cancellableRace()` method and the
returned cancel method now takes a cancellation message.