feat: RaceCancellation PromiseLike support
Last update I removed the ability to pass in an
existing promise to RaceCancellation and this adds
it back.

The assumption of this library is you don't create
promises you are not ready to make part of the
current chain. So this was removed to simplify
code but the concept of not making promises before
you place them into the chain seems like a tough
hurdle for people to understand.

When people close over a promise to do
`await raceCancellation(() => promise)` when it is
already in the cancelled state, it will not invoke
the passed in task and create the `Promise.race`
which is a feature of this library. The issue with
this behavior, is if the promise isn't created by
the task but closed over, it has the potential to
trigger an unhandled promise rejection in this
scenario.

This change will always create the race regardless
of the cancellation state if you pass in a promise
like instead of a function which allows you to opt
out of the short circuiting.